update integer to long

+++ src/main/java/safide/erp/inventory/application/port/InveProdCataService.java
@@ -4,7 +4,7 @@ import safide.erp.generic.application.port.GeneErpService;
 import safide.erp.inventory.domain.model.InveProdCata;
 import safide.erp.inventory.domain.port.IInveProdCataRepository;
 
-public class InveProdCataService extends GeneErpService<InveProdCata, Integer> {
+public class InveProdCataService extends GeneErpService<InveProdCata, Long> {
 	public InveProdCataService(IInveProdCataRepository iInveProdCataRepository) {
 		super(iInveProdCataRepository);	}
