+++ src/main/java/safide/erp/inventory/application/port/InveProdHeadService.java
@@ -4,7 +4,7 @@ import safide.erp.generic.application.port.GeneErpService;
 import safide.erp.inventory.domain.model.InveProdHead;
 import safide.erp.inventory.domain.port.IInveProdHeadRepository;
 
-public class InveProdHeadService extends GeneErpService<InveProdHead, Integer> {
+public class InveProdHeadService extends GeneErpService<InveProdHead, Long> {
 	public InveProdHeadService(IInveProdHeadRepository iInveProdHeadRepository) {
 		super(iInveProdHeadRepository);
 	}
