+++ src/main/java/safide/erp/inventory/domain/port/IInveProdHeadRepository.java
@@ -3,5 +3,5 @@ package safide.erp.inventory.domain.port;
 import safide.erp.generic.domain.port.IGeneErpRepository;
 import safide.erp.inventory.domain.model.InveProdHead;
 
-public interface IInveProdHeadRepository extends IGeneErpRepository<InveProdHead, Integer> {
+public interface IInveProdHeadRepository extends IGeneErpRepository<InveProdHead, Long> {
 }
