+++ src/main/java/safide/erp/inventory/infrastructure/adapter/input/IInveProdHeadCrudRepository.java
@@ -3,6 +3,6 @@ package safide.erp.inventory.infrastructure.adapter.input;
 import org.springframework.data.repository.CrudRepository;
 import safide.erp.inventory.infrastructure.entity.InveProdHeadEntity;
 
-public interface IInveProdHeadCrudRepository extends CrudRepository<InveProdHeadEntity, Integer> {
+public interface IInveProdHeadCrudRepository extends CrudRepository<InveProdHeadEntity, Long> {
 
 }
