+++ src/main/java/safide/erp/inventory/infrastructure/entity/InveProdHeadEntity.java
@@ -20,7 +20,7 @@ public class InveProdHeadEntity {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "id_inprhe")
-	private Integer id;
+	private Long id;
 	@Column(unique = true, nullable = true)
 	private String name_inprhe;
 	private String descrip_inprhe;
@@ -39,7 +39,7 @@ public class InveProdHeadEntity {
     private Integer idrecur_inprhe;
     private String brand_inprhe;
     private BigDecimal stock_inprhe;
-    private Integer id_inprca;
-    private Integer id_macohe;
-    private Integer id_macode;
+    private Long id_inprca;
+    private Long id_macohe;
+    private Long id_macode;
 }
