+++ src/main/java/safide/erp/sale/domain/port/ISaleCustHeadRepository.java
@@ -3,5 +3,5 @@ package safide.erp.sale.domain.port;
 import safide.erp.generic.domain.port.IGeneErpRepository;
 import safide.erp.sale.domain.model.SaleCustHead;
 
-public interface ISaleCustHeadRepository extends IGeneErpRepository<SaleCustHead, Integer> {
+public interface ISaleCustHeadRepository extends IGeneErpRepository<SaleCustHead, Long> {
 }
