+++ src/main/java/safide/erp/system/infrastructure/adapter/input/ISystUserHeadCrudRepository.java
@@ -3,6 +3,6 @@ package safide.erp.system.infrastructure.adapter.input;
 import org.springframework.data.repository.CrudRepository;
 import safide.erp.system.infrastructure.entity.SystUserHeadEntity;
 
-public interface ISystUserHeadCrudRepository extends CrudRepository<SystUserHeadEntity, Integer> {
+public interface ISystUserHeadCrudRepository extends CrudRepository<SystUserHeadEntity, Long> {
 
 }
