+++ src/main/java/safide/erp/system/infrastructure/entity/SystUserHeadEntity.java
@@ -22,7 +22,7 @@ public class SystUserHeadEntity {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "id_syushe")
-	private Integer id;
+	private Long id;
 	@Column(unique = true, nullable = true)
     private String name_syushe;
     private String delegat_syushe;
@@ -31,7 +31,7 @@ public class SystUserHeadEntity {
     private LocalDateTime datacre_syushe;
     @Enumerated(EnumType.STRING)
     private SystUserType sysusertype;
-    private Integer id_macohe;
-    private Integer id_macode;
+    private Long id_macohe;
+    private Long id_macode;
 
 }
